+++ app/helpers/dislikes_helper.rb
@@ -1,0 +1,2 @@
+module DislikesHelper
+end
