+++ config/routes.rb
@@ -15,7 +15,8 @@ Rails.application.routes.draw do
   resources :users
   resources :account_activations, only: [:edit]
   resources :password_resets,     only: [:new, :create, :edit, :update]
-  resources :products, only: [:index, :show]
+  resources :products
+  resources :dislikes, only: [:create, :destroy]
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
   
 end
